requirements.txt
streamlit
matplotlib
plotly
pandas
numpy
ipython
mpld3
Pillow